+++ cmd/preview.go
@@ -34,6 +34,7 @@ func newPreviewCmd() *cobra.Command {
 	// Flags for engine.UpdateOptions.
 	var analyzers []string
 	var diffDisplay bool
+	var jsonDisplay bool
 	var parallel int
 	var showConfig bool
 	var showReplacementSteps bool
@@ -72,6 +73,7 @@ func newPreviewCmd() *cobra.Command {
 					SuppressOutputs:      suppressOutputs,
 					IsInteractive:        cmdutil.Interactive(),
 					DiffDisplay:          diffDisplay,
+					JSONDisplay:          jsonDisplay,
 					Debug:                debug,
 				},
 			}
@@ -134,6 +136,9 @@ func newPreviewCmd() *cobra.Command {
 	cmd.PersistentFlags().BoolVar(
 		&diffDisplay, "diff", false,
 		"Display operation as a rich diff showing the overall change")
+	cmd.Flags().BoolVarP(
+		&jsonDisplay, "json", "j", false,
+		"Serialize the preview diffs, operations, and overall output as JSON")
 	cmd.PersistentFlags().IntVarP(
 		&parallel, "parallel", "p", defaultParallel,
 		"Allow P resource operations to run in parallel at once (1 for no parallelism). Defaults to unbounded.")