+++ pkg/backend/display/options.go
@@ -23,8 +23,9 @@ type Options struct {
 	ShowReplacementSteps bool                // true to show the replacement steps in the plan.
 	ShowSameResources    bool                // true to show the resources that aren't updated in addition to updates.
 	SuppressOutputs      bool                // true to suppress output summarization, e.g. if contains sensitive info.
-	SummaryDiff          bool                // If the diff display should be summarized
-	IsInteractive        bool                // If we should display things interactively
-	DiffDisplay          bool                // true if we should display things as a rich diff
+	SummaryDiff          bool                // true if diff display should be summarized.
+	IsInteractive        bool                // true if we should display things interactively.
+	DiffDisplay          bool                // true if we should display things as a rich diff.
+	JSONDisplay          bool                // true if we should emit the entire diff as JSON.
 	Debug                bool                // true to enable debug output.
 }